move _create_full_field_tiff into a utils file
