+++ tests/modules/mesoscope_splitting/__init__.py
@@ -1,0 +1,1 @@
+# empty
